generic_ui/requirements.txt
numpy
pandas
PySide2
PyInstaller
pyqtgraph
pyserial
pyusb

